Fixed the ID again
: (

+++ Backend/APIBanca/Model/DTO/EmpresaDTO.cs
@@ -9,7 +9,7 @@ namespace Model.DTO
     public class EmpresaDTO
     {
         public string cuit { get; set; }
-        public long id_account { get; set; }
+        public int id_account { get; set; }
         public string phone { get; set; }
         public string address { get; set; }
     }
